Integrate jiayev/pp at 3537b166d with the Vulkan presentation path, HDR color management, and Streamline history resets. Resolve imported PP conflicts and duplicate translations.

Update the CS-side CER LinearLightingSettings layout to 64 bytes and upload managed fog colors. CER must synchronize this interface and the new working-space color handling before use.

+++ features/Post Processing/Shaders/PostProcessing/LensFlare/lensflare.cs.hlsl
@@ -57,9 +57,9 @@ cbuffer LensFlareConstants : register(b1)
 // Utilities
 // ============================================================
 
-float GetGhostScale(int i)
+float GetGhostScale(uint i)
 {
-	return GhostScalesPacked[i / 4][i % 4];
+	return GhostScalesPacked[i / 4u][i % 4u];
 }
 
 // Fisheye UV distortion (based on Shadertoy by Crucifer)
@@ -138,7 +138,7 @@ float DiscMask(float2 screenPos)
 	[branch] if (GhostStrength > EPSILON)
 	{
 		// Chromatic aberration on input for ghosts
-		for (int i = 0; i < NUM_GHOSTS; i++) {
+		for (uint i = 0; i < (uint)NUM_GHOSTS; i++) {
 			float4 ghostColor = GhostColors[i];
 			float ghostScale = GetGhostScale(i);
